+++ app/src/main/java/com/aircalc/converter/di/AppModule.kt
@@ -58,7 +58,6 @@ object AppModule {
     }
 
     @Provides
-    @Singleton
     fun provideTimerManager(): TimerManager {
         return TimerManager()
     }